ath10k: Don't request OFDM IE when there are no OFDM rates.
This makes the IE's in the ProbeRequests not advertise rates
that the vif cannot use.

Also, add some CT-firmware only flags that allow disabling
HT and VHT ies.

As an aside, at least for CCK and OFDM rates, the mac80211
stack is already creating IEs for this, so probe requests
end up with duplicate entries.  This issue is not addressed
in this patch.

Signed-off-by: Ben Greear <greearb@candelatech.com>