arm64: dts: mediatek: mt7988a-bpi-r4: allow hw variants of bpi-r4
Sinovoip has released other variants of Bananapi-R4 board.
The known changes affecting only the LAN SFP+ slot which is replaced
by a 2.5G phy with optional PoE.

As the SFP-Ports are not upstreamed yet this patch simply renames the
current board dts to dtsi and creating new dts for the board including
the new dtsi.

Signed-off-by: Frank Wunderlich <frank-w@public-files.de>