[FOR-UPSTREAM] uio: uio_pdrv_genirq: remove incorrect MODULE_DEVICE_TABLE
This was harmless albeit useless when it was added, but the placeholder
table entry added later causes it to get the wildcard modaliases
of:N*T* and of:N*T*C*, which match every possible node.

This means that it is unnecessarily loaded on any machine that
probes for a module for a device tree node with no assigned driver,
and silently produces bizarre results for modprobe(8) and modinfo(8)
commands that should properly fail.

Fixes: 05c3e0bb5629b897b0459e4bfb1b93d729033b99 ("UIO: allow binding uio_pdrv_genirq.c to devices using command line option")
Signed-off-by: Emily <hello@emily.moe>